+++ src/Installer/InstallerFactory.php
@@ -74,12 +74,23 @@ class InstallerFactory
     public $noInit;
 
     /**
+     * @var bool
+     */
+    public $initonly;
+
+    /**
      * Given a big bag of install options, add installers to the collection.
      *
      * @param InstallerCollection $installers Installers will be added to this
      */
     public function addInstallers(InstallerCollection $installers)
     {
+        if ($this->initonly && ($this->plugin->hasBehatFeatures() || $this->plugin->hasUnitTests())) {
+            $installers->add(new TestSuiteInstaller($this->moodle, $this->plugin, $this->execute));
+
+            return;
+        }
+
         $installers->add(new MoodleInstaller($this->execute, $this->database, $this->moodle, new MoodleConfig(), $this->repo, $this->branch, $this->dataDir));
         $installers->add(new PluginInstaller($this->moodle, $this->plugin, $this->pluginsDir, $this->dumper));
         $installers->add(new VendorInstaller($this->moodle, $this->plugin, $this->execute));